+++ buckt_test.go
@@ -636,7 +636,7 @@ func TestResolveBackend(t *testing.T) {
 			Source:           source,
 			Target:           target,
 		}
-		result := resolveBackend(mediaDir, bc, mockLogger, mockLRU, nil)
+		result := resolveBackend(mediaDir, bc, mockLogger, mockLRU, nil, nil)
 		_, ok := result.(*backend.MigrationBackendService)
 		assert.True(t, ok)
 	})
@@ -646,7 +646,7 @@ func TestResolveBackend(t *testing.T) {
 		bc := BackendConfig{
 			Source: source,
 		}
-		result := resolveBackend(mediaDir, bc, mockLogger, mockLRU, nil)
+		result := resolveBackend(mediaDir, bc, mockLogger, mockLRU, nil, nil)
 		// Placeholder should be replaced with real local backend
 		_, ok := result.(*backend.LocalFileSystemService)
 		assert.True(t, ok)
@@ -657,7 +657,7 @@ func TestResolveBackend(t *testing.T) {
 		bc := BackendConfig{
 			Source: source,
 		}
-		result := resolveBackend(mediaDir, bc, mockLogger, mockLRU, nil)
+		result := resolveBackend(mediaDir, bc, mockLogger, mockLRU, nil, nil)
 		assert.Equal(t, source, result)
 	})
 
@@ -666,14 +666,14 @@ func TestResolveBackend(t *testing.T) {
 		bc := BackendConfig{
 			Target: target,
 		}
-		result := resolveBackend(mediaDir, bc, mockLogger, mockLRU, nil)
+		result := resolveBackend(mediaDir, bc, mockLogger, mockLRU, nil, nil)
 		_, ok := result.(*backend.LocalFileSystemService)
 		assert.True(t, ok)
 	})
 
 	t.Run("No Source or Target", func(t *testing.T) {
 		bc := BackendConfig{}
-		result := resolveBackend(mediaDir, bc, mockLogger, mockLRU, nil)
+		result := resolveBackend(mediaDir, bc, mockLogger, mockLRU, nil, nil)
 		_, ok := result.(*backend.LocalFileSystemService)
 		assert.True(t, ok)
 	})